requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
python-dotenv==1.0.1
pandas==2.1.4
openpyxl==3.1.2
